move response file to response dir
