+++ lib/alexa_ruby.rb
@@ -14,7 +14,7 @@ require 'alexa_ruby/request/launch_request'
 require 'alexa_ruby/request/intent_request'
 require 'alexa_ruby/request/intent_request/slot'
 require 'alexa_ruby/request/session_ended_request'
-require 'alexa_ruby/response'
+require 'alexa_ruby/response/response'
 require 'alexa_ruby/response/audio_player'
 require 'alexa_ruby/response/card'
 require 'alexa_ruby/version'
